Better redisplay routine in game loop
By letting Emacs to decide whether to redraw the buffer or not by simply pausing
the current loop (`run-with-time`) can lead to a lower frame rate than what's
possible.

NOTE: we measure the time between loop iterations, we don't know if between
those iterations Emacs has redraw the buffer or not.

So better to "force" Emacs to redraw the buffer and make a smaller pause between
the loops (`0.001` instead of `0.025`).

NOTE: the pause (`0.001`) is needed to let Emacs to collect inputs.